The login function is working :)
